feat(utils): separate interfaces
