+++ galasa-ui/src/utils/interfaces/testRuns.ts
@@ -1,0 +1,45 @@
+/*
+ * Copyright contributors to the Galasa project
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+import { AmPm } from './common';
+
+/**
+ * Interfaces related to test runs, filtering, and time-based operations
+ */
+
+export interface RunMetadata {
+  runId: string;
+  result: string;
+  status: string;
+  testName: string;
+  runName: string;
+  bundle: string;
+  submissionId: string;
+  group: string;
+  requestor: string;
+  submitted: string;
+  startedAt: string;
+  finishedAt: string;
+  duration: string;
+  tags: string[];
+}
+
+export interface TimeFrameValues {
+  fromDate: Date;
+  fromTime: string;
+  fromAmPm: AmPm;
+  toDate: Date;
+  toTime: string;
+  toAmPm: AmPm;
+  durationDays: number;
+  durationHours: number;
+  durationMinutes: number;
+}
+
+export interface RunLog {
+  content: string;
+  route: string;
+}
